Update train.py
Edit train.py for this Machine learning model known as a convolutional neural network. It was quickly trained thanks to transfer learning, which takes advantage of an existing trained model, one that was trained on a larger dataset and with more powerful compute. training of this model was limited to under half an hour and acheived 95% accuracy on the test set, without overfitting, as the model performed well when deployed on a web application. Written by Brian lesko, robotics engineer and mechanical engineering masters graduate.